Refactor: 패키지명 수정 및 h2 db 연결 경로 수정
