--- src/main/java/com/codessquad/domain/question/Question.java
@@ -1,68 +1,0 @@
-package com.codessquad.domain.question;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-
-@Entity
-public class Question {
-
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    @Column(nullable = false, length = 30)
-    private String writer;
-
-    @Column(nullable = false)
-    private String title;
-
-    @Column(nullable = false)
-    private String contents;
-
-    private LocalDateTime dateTime;
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getWriter() {
-        return writer;
-    }
-
-    public void setWriter(String writer) {
-        this.writer = writer;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getContents() {
-        return contents;
-    }
-
-    public void setContents(String contents) {
-        this.contents = contents;
-    }
-
-    public void setDateTime(LocalDateTime dateTime) {
-        this.dateTime = dateTime;
-    }
-
-    public String getFormattedDateTime() {
-        return dateTime.format(DateTimeFormatter.ofPattern("YYYY-MM-dd HH:mm:ss"));
-    }
-
-    public void update(Question updateQuestion) {
-        this.title = updateQuestion.title;
-        this.contents = updateQuestion.contents;
-    }
-}
