+++ src/main/resources/application.properties
@@ -12,7 +12,7 @@ spring.h2.console.enabled=true
 spring.h2.console.path=/h2-console
 
 #DB Connection 설정
-spring.datasource.url=jdbc:h2:~/codesquad/workspace/myslipp;DB_CLOSE_ON_EXIT=FALSE
+spring.datasource.url=jdbc:h2:~/codesquad/workspace/java-qna;DB_CLOSE_ON_EXIT=FALSE
 spring.datasource.driverClassName=org.h2.Driver
 #JPA 데이터베이스 플랫폼을 지정
 spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
